Split folding ranges into editor and lsp parts
